scanner/requirements.txt
flask==1.1.2
gunicorn==20.1.0
jinja2==2.11.3
markupsafe==2.0.1
werkzeug==2.0.3
itsdangerous==1.1.0
PyPDF2==2.10.5
python-docx==0.8.11
psycopg2==2.9.3
python-dotenv==0.21.0
py-tlsh==4.7.2
